map-widget: fix chronology of QML class registration
The QML types needs to be registered before the MainWindow
instance is created on the desktop version, otherwise
the MapWidget instance will be created and it will fail
with a missing QML namespace.

In subsurface-helper.cpp, move register_qml_types() from run_ui()
to init_ui(), as later in init_ui(), MainWindow is instantiated
for the first time in the desktop version.

Ref #1500

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>