Cleanup: rename dive_getUniqID() and handle double ids gracefully
dive_getUniqID() actually *set* the unique id of a dive. Therefore,
rename this function to dive_setUniqID(). Weirdly, all callers of the
function used the pattern
	dive.id = dive_getUniqID(&dive);
so the dive-id was assigned twice, once inside and once outside the
function. Change this to
	dive_setUniqId(&dive)
and don't return the id. It can be read from dive.id anyway.

Finally, fix a small inconsistency:
If a dive already had an id, this id would have been registered and
a debug message written. Creation of an already-registered id would
on the other hand raise an assert(). But the latter can only happen
as a consequence of the former - so why be once graceful and once
terminate? Therefore, make both cases graceful.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>