Add performance test case
This requires the user to manually copy the large test file into the
dives directory (because I really don't want to add this to the repo);
instructions how to do that are displayed if the file is missing.

Next it uses the git version of that same file (but prefetches it to
try and remove the network speed from what is being measured).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>