Build: suppress inconsistent-missing-override warnings on clang
In a previous commit all override-modifiers were thrown out owing
to warning floods caused by the inconsistent-missing-override
flag on clang.

Re-adding overrides is now very painful because it's an all-or-nothing
thing. Let's disable the warning for now. If we reastablished good
coverage with override, we might think about re-enabling the warning.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>